Added base controller for api module
Created a module to produce JSON api to be used by R (and possibly other
things in the future), with a base module to generalize request types and
pagination e.g. #GET api/{plural_resource_name}/1 for page 1 of the index
of a resource with a given resource name.

See https://codelation.com/blog/rails-restful-api-just-add-water for more
information